backend/requirements.txt
# Doubao Image Generation Backend

fastapi==0.109.0
uvicorn[standard]==0.27.0
openai>=1.56.0
python-dotenv==1.0.0
pydantic==2.6.0
pydantic-settings==2.1.0
pillow==10.2.0
httpx==0.26.0
aiofiles==23.2.1
python-multipart==0.0.6
